Distro: move `rhel8` to `distro/rhel` directory
