+++ pkg/distro/rhel/rhel8/ami.go
