+++ pkg/distro/rhel/rhel8/azure.go
