+++ pkg/distro/rhel/rhel8/bare_metal.go
