+++ pkg/distro/rhel/rhel8/gce.go
