+++ pkg/distro/rhel/rhel8/images.go
