+++ pkg/distro/rhel/rhel8/imagetype.go
