+++ pkg/distro/rhel/rhel8/partition_tables.go
