+++ pkg/distro/rhel/rhel8/qcow2.go
