+++ pkg/distro/rhel/rhel8/vmdk.go
