+++ pkg/distro/rhel/rhel8/workloads.go
