+++ pkg/distrofactory/distrofactory.go
@@ -7,9 +7,9 @@ import (
 	"github.com/osbuild/images/pkg/distro"
 	"github.com/osbuild/images/pkg/distro/fedora"
 	"github.com/osbuild/images/pkg/distro/rhel/rhel10"
+	"github.com/osbuild/images/pkg/distro/rhel/rhel8"
 	"github.com/osbuild/images/pkg/distro/rhel/rhel9"
 	"github.com/osbuild/images/pkg/distro/rhel7"
-	"github.com/osbuild/images/pkg/distro/rhel8"
 	"github.com/osbuild/images/pkg/distro/test_distro"
 )
