+++ pkg/distroidparser/idparser.go
@@ -4,9 +4,9 @@ import (
 	"github.com/osbuild/images/pkg/distro"
 	"github.com/osbuild/images/pkg/distro/fedora"
 	"github.com/osbuild/images/pkg/distro/rhel/rhel10"
+	"github.com/osbuild/images/pkg/distro/rhel/rhel8"
 	"github.com/osbuild/images/pkg/distro/rhel/rhel9"
 	"github.com/osbuild/images/pkg/distro/rhel7"
-	"github.com/osbuild/images/pkg/distro/rhel8"
 )
 
 var DefaultParser = NewDefaultParser()
